Refactored NamedLock.withLock
- Trying to also fix dart_project_test.dart
