+++ dcli/test/src/util/named_lock_test.dart
@@ -14,55 +14,23 @@ import 'dart:io';
 import 'dart:isolate';
 
 import 'package:async/async.dart';
-import 'package:dcli/dcli.dart' hide NamedLock;
+import 'package:dcli/dcli.dart';
 import 'package:dcli_core/dcli_core.dart' as core;
 import 'package:path/path.dart' hide equals;
-import 'package:runtime_named_locks/runtime_named_locks.dart' show ExecutionCall, NamedLock;
 import 'package:test/test.dart';
 
-// const port = 9003;
-
 void main() {
-  test('lock path', () {
-    final lockPath = join(rootPath, Directory.systemTemp.path, 'dcli', 'locks');
-    print(lockPath);
-  });
 
-  test('unsafe exception catch', () async {
+  test('exception catch', () async {
+
     expect(
-      () {
-        NamedLock.guard<void, DCliException>(
-          name: 'unsafe-exception-lock',
-          execution: ExecutionCall<void, DCliException>(callable: () {
-            throw DCliException('fake exception');
-          }),
-        );
-      },
-      throwsA(isA<DCliException>()),
-    );
-  });
-
-  test('safe exception catch', () async {
-    final call = NamedLock.guard<void, DCliException>(
-      name: 'safe-exception-lock',
-      execution: ExecutionCall(
-        callable: () {
-          throw DCliException('fake exception');
-        },
-        safe: true,
-      ),
-    );
-
-    expect(call.successful.get, false);
-    expect(call.error.get?.anticipated.get, isA<DCliException>());
-    expect(
-      () => call.error.get?.rethrow_(),
-      throwsA(isA<DCliException>()),
+       NamedLock(suffix: 'exception').withLock(() => throw DCliException('fake exception')),
+       throwsA(isA<DCliException>()),
     );
   });
 
   test(
-    'File Operations with NamedLock.guard',
+    'withLock',
     () async {
       await core.withTempDirAsync(
         (fs) async {
@@ -78,7 +46,7 @@ void main() {
             await portMid.first;
             await portFore.first;
 
-            print('reading logfile');
+            print('readling logfile');
 
             final actual = read(logFile).toList();
 
@@ -104,10 +72,11 @@ void main() {
         keep: true,
       );
     },
+    skip: false,
   );
 
   test(
-    'Thrash Test',
+    'Thrash test',
     () async {
       Settings().setVerbose(enabled: false);
       if (exists(_lockCheckPath)) {
@@ -133,12 +102,12 @@ void main() {
       expect(exists(_lockFailedPath), equals(false));
     },
     timeout: const Timeout(Duration(minutes: 30)),
-    skip: false,
   );
 }
 
 Future<ReceivePort> spawn(String message, String logFile) async {
-  final back = await Isolate.spawn(writeToLog, '$message;$logFile', paused: true);
+  final back =
+      await Isolate.spawn(writeToLog, '$message;$logFile', paused: true);
   final port = ReceivePort();
   back
     ..addOnExitListener(port.sendPort)
@@ -146,23 +115,19 @@ Future<ReceivePort> spawn(String message, String logFile) async {
   return port;
 }
 
-void writeToLog(String data) {
+FutureOr<void> writeToLog(String data) {
   final parts = data.split(';');
   final message = parts[0];
   final log = parts[1];
-
-  NamedLock.guard(
-    name: 'test-log-lock',
-    execution: ExecutionCall(callable: () {
-      var count = 0;
-      for (var i = 0; i < 4; i++) {
-        final l = '$message + ${count++}';
-        print(l);
-        log.append(l);
-        sleep(1);
-      }
-    }),
-  );
+  NamedLock(suffix: 'test.lock').withLock(()  {
+    var count = 0;
+    for (var i = 0; i < 4; i++) {
+      final l = '$message + ${count++}';
+      print(l);
+      log.append(l);
+      sleep(1);
+    }
+  });
 
   print('Finished Write to Log for $message');
 }
@@ -171,33 +136,28 @@ const _lockCheckPath = '/tmp/lockcheck';
 final _lockFailedPath = join(_lockCheckPath, 'lock_failed');
 
 /// must be a global function as we us it to spawn an isolate
-Future<void> worker(int instance) async {
+FutureOr<void> worker(int instance) {
   Settings().setVerbose(enabled: false);
   print('starting worker instance $instance ${DateTime.now()}');
+  NamedLock(suffix: 'gshared-compile').withLock(() {
+    print('acquired lock worker $instance  ${DateTime.now()}');
+    final inLockPath = join(_lockCheckPath, 'inlock');
 
-  NamedLock.guard(
-    name: 'gshared-compile-lock',
-    execution: ExecutionCall(callable: () {
-      print('acquired lock worker $instance  ${DateTime.now()}');
-      final inLockPath = join(_lockCheckPath, 'inlock');
+    /// If the [inLockPath] file exists
+    /// then the lock has been breached.
+    if (exists(inLockPath)) {
+      touch(_lockFailedPath, create: true);
+      throw DCliException(
+        'NamedLock for $instance failed as another lock is active',
+      );
+    }
 
-      /// If the [inLockPath] file exists
-      /// then the lock has been breached.
-      if (exists(inLockPath)) {
-        touch(_lockFailedPath, create: true);
-        throw DCliException(
-          'NamedLock for $instance failed as another lock is active',
-        );
-      }
+    touch(inLockPath, create: true);
 
-      touch(inLockPath, create: true);
-
-      sleep(2);
-      print('finished work $instance  ${DateTime.now()}');
-      delete(inLockPath);
-    }),
-  );
-
+    sleep(2);
+    print('finished work $instance  ${DateTime.now()}');
+    delete(inLockPath);
+  });
   print('released lock $instance  ${DateTime.now()}');
 }
